+++ src/Dialog/Dialog.tsx
@@ -1,4 +1,4 @@
-import React, {useCallback, useEffect, useRef} from 'react'
+import React, {useCallback, useEffect, useRef, useState} from 'react'
 import styled from 'styled-components'
 import Button, {ButtonPrimary, ButtonDanger, ButtonProps} from '../Button'
 import Flex from '../Flex'
@@ -341,11 +341,16 @@ const buttonTypes = {
 const Buttons: React.FC<{buttons: DialogButtonProps[]}> = ({buttons}) => {
   const autoFocusRef = useRef<HTMLButtonElement>(null)
   let autoFocusCount = 0
+  const [hasRendered, setHasRendered] = useState(0)
   useEffect(() => {
-    if (autoFocusRef.current) {
-      autoFocusRef.current.focus()
+    // hack to work around dialogs originating from other focus traps.
+    if (hasRendered === 1) {
+      autoFocusRef.current?.focus()
+    } else {
+      setHasRendered(hasRendered + 1)
     }
-  }, [])
+  }, [hasRendered])
+
   return (
     <>
       {buttons.map((dialogButtonProps, index) => {