+++ src/ThemeProvider.tsx
@@ -90,7 +90,7 @@ export function useTheme() {
   return React.useContext(ThemeContext)
 }
 
-export function useColorSchemeVar(values: Partial<Record<string, string>>, fallback?: string) {
+export function useColorSchemeVar(values: Partial<Record<string, string>>, fallback: string) {
   const {colorScheme = ''} = useTheme()
   return values[colorScheme] ?? fallback
 }